pyproject.toml
[tool.poetry]
name = "publishgpt"
version = "0.0.1"
description = "Package to create Books with ChatGPT"
authors = ["markus <datamastery87@gmail.com>"]
license = "MIT"


[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.27.7"
pypandoc = "^1.11"
requests = "^2.30.0"
markdown = "^3.4.3"
tqdm = "^4.65.0"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
publishgpt = "publishgpt.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
